rtps test: Showcase an app that uses UART rtps stuff
This app is *not* meant to be here - just to simplify code sharing.

It's the other side of examples/rtps: reads from UART and publishes via
RTPS, subscribe via RTPS and writes to UART.

Following examples/rtps, it only publishes vehicle_status,
sensors_combined and vehicle_command_ack. Subscribe to vehicle_command.

This example runs on Aero, in the same way mavlink-router there.