rtps: Exit when receive shutdown command
As uORB shutdown command is low prio, if rtps keeps running, it risks
not ever allowing shutdown command to run - for instance, if it keeps
receiving messages.

This was the reason why it was necessary a flush and sleep right after
RTPS shutdown command was sent by Tools/px_uploader. With rtps side
fixed, no need for this hack anymore.