rtps-router: Companion computer side of rtps bridge
*This program doesn't belong to this repo. Here to ease code sharing*

Reads MAVLink and RTPS messages from UART, publish RTPS ones and send
MAVLink ones via UDP to mavlink-router.

This app is still a PoC to demostrate companion computer side of RTPS
effort.

To build it, create a build dir inside rtps-router directory and call
cmake:

    $ mkdir build && cd build
    $ cmake ..
    $ make

An executable called `rtps_router` shall be created. To run it on Intel
Aero:

    $ ./rtps_router /dev/ttyS1 460800

Note that eProsima libraries 'libfastrtps.so' and 'libfastcdr.so' should
be on libraries path (for instance, LD_LIBRARY_PATH).